Factors out the CPC's simple FDC adaptor.
